pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "smart-tool-select"
version = "0.1.0"
description = "Lightweight semantic tool selector for LLM tool binding"
readme = "README.md"
license = { file = "LICENSE" }
license-files = ["LICENSE"]
requires-python = ">=3.9"
dependencies = [
  "sentence-transformers>=2.2.2",
  "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7.4"]
hnswlib = ["hnswlib>=0.7.0"]
yaml = ["pyyaml>=6.0.1"]

[tool.hatch.build.targets.wheel]
packages = ["app/smart_tool_select"]
include = ["app/smart_tool_select/models/**", "LICENSE"]

[tool.hatch.build.targets.sdist]
include = ["app/smart_tool_select", "app/smart_tool_select/models/**", "README.md", "LICENSE", "pyproject.toml", "scripts/vendor_model.py"]
